Correct rdfs:subClassOf for gs1de:ImportClassification and gs1de:ContainedRegulatedCommodityDetails
These are definitely not rdfs:subClassOf gs1:Product because they don't inherit the properties of gs1:Product in the way that true subclasses such as gs1:WearableProduct or gs1:FoodBeverageTobaccoProduct do.

I suspect that someone in the team has been confused by the tendency of the GDSN data model to use the term 'parentClass' when they should really be using rdfs:domain if they wanted to align correctly with W3C Linked Data standards.  However. parentClass in the GDSN data model usually corresponds to the rdfs:domain rather than indicating a subclass relationship.

For these two classes, you already have properties gs1de:importClassification and gs1de:containedRegulatedCommodities that express a rdfs:range relationship to these classes - and that's sufficient to ensure that they're correctly connected in the data model.